pytest: Migrate all remaining tests into test_misc.py
